Create email and password components
